Operators - 1 we will do the sum of int from 0 to 3 included
Operators - 2
same as Operators - 1 but using the operator +=

Operators - 3
we will substract from result values from 0 to 2 included so it'll go to 0

Operators - 4
same as Operators - 3 but using the operator -=

Operators - 5
we will calculate the rest between 11 and 5 which is 1

Operators - 6
it is the answer to everything in the universe and you can find it with a calculator

Operators - 7
the && are calculated before the || so we will have
e1=true && true=true
e2=false && false=false
e3=e1||e2=true

Operators - 8
same explanation as 7

e1=false&&true=false
e2=true&&true=true
e3=e1||e2

Operators - 9

a new element will be pushed in the result everytime log is called. first when we initialise i, second when we check the condition. Third when we call it with body. Fourth with the incrementation and firth when we check the condition again