Truthyness
isTruthy return true if the parameter is true else it return false.

when use a single integer in a condition it will always be true unless it's 0, This one is always false.

null and undefined are always false

In 6 it's true because we Boolean will return true if the Boolean have been correctly created

In 7 we verify that the value of the boolean is false

In 8 we will true because false is always equal to itself